+++ .gitattributes
@@ -39,3 +39,5 @@ Makefile text
 *.zip binary diff=hex
 *.dmg binary diff=hex
 *.rpi-sdcard binary diff=hex
+*.foo binary diff=hex
+xz-without-extension binary diff=hex
