+++ lib/image-stream/index.js
@@ -20,7 +20,7 @@ const _ = require('lodash');
 const Bluebird = require('bluebird');
 const fs = Bluebird.promisifyAll(require('fs'));
 const stream = require('stream');
-const utils = require('./utils');
+const mime = require('./mime');
 const handlers = require('./handlers');
 const supportedFileTypes = require('./supported');
 const errors = require('../shared/errors');
@@ -76,9 +76,8 @@ exports.getFromFilePath = (file) => {
       });
     }
 
-    return utils.getArchiveMimeType(file).then((type) => {
-      const MIME_TYPE_RAW_IMAGE = 'application/octet-stream';
-      const mimeType = _.has(handlers, type) ? type : MIME_TYPE_RAW_IMAGE;
+    return mime.getMimeTypeFromFileName(file).then((type) => {
+      const mimeType = _.has(handlers, type) ? type : mime.DEFAULT_MIME_TYPE;
       return _.invoke(handlers, mimeType, file, {
         size: fileStats.size
       });
