+++ lib/image-stream/utils.js
@@ -16,48 +16,7 @@
 
 'use strict';
 
-const _ = require('lodash');
 const Bluebird = require('bluebird');
-const fileType = require('file-type');
-const mime = require('mime-types');
-const fs = require('fs');
-
-/**
- * @summary Get archive mime type
- * @function
- * @public
- *
- * @param {String} filename - file path
- * @fulfil {String} - mime type
- * @returns {Promise}
- *
- * @example
- * utils.getArchiveMimeType('path/to/raspberrypi.img.gz').then((mimeType) => {
- *   console.log(mimeType);
- * });
- */
-exports.getArchiveMimeType = (filename) => {
-  const MIME_TYPE_RAW_IMAGE = 'application/octet-stream';
-  const mimeType = mime.lookup(filename);
-
-  if (mimeType) {
-    return Bluebird.resolve(mimeType);
-  }
-
-  const FILE_TYPE_ID_BYTES = 262;
-
-  return Bluebird.using(fs.openAsync(filename, 'r').disposer((fileDescriptor) => {
-    return fs.closeAsync(fileDescriptor);
-  }), (fileDescriptor) => {
-    const BUFFER_START = 0;
-    const buffer = Buffer.alloc(FILE_TYPE_ID_BYTES);
-
-    return fs.readAsync(fileDescriptor, buffer, BUFFER_START, FILE_TYPE_ID_BYTES, null).then(() => {
-      return _.get(fileType(buffer), [ 'mime' ], MIME_TYPE_RAW_IMAGE);
-    });
-  });
-
-};
 
 /**
  * @summary Extract the data of a readable stream
